Move test of ap201 to ruststep/tests
